[refactoring] Categories, Regions를 리팩토링한다.
범용적으로 사용이 가능할 것 같아 Buttons 컴포넌트로 통일합니다
